WIP: Started to populate util_portability
